chore(directory): reorganize project structure for clearer separation of concerns
- move constants and messages to config/
- move input and output handlers to view/
- separate domain into model/ and service/ for better modularity
- adjust package structure to improve scalability and maintainability
