+++ src/main/kotlin/lotto/config/LottoConstants.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.config
 
 object LottoConstants {
     const val LOTTO_PRICE = 1_000
