+++ src/main/kotlin/lotto/config/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.config
 
 interface NumberGenerator {
     fun generate() : List<Int>
