+++ src/main/kotlin/lotto/domain/model/Lotties.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.model
 
 class Lotties(val value: List<Lotto>) {
 }
