+++ src/main/kotlin/lotto/domain/model/Rank.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.model
 
 enum class Rank(
     val matchCount: Int,
