+++ src/main/kotlin/lotto/domain/model/WinningNumbers.kt
@@ -1,3 +1,3 @@
-package lotto.domain
+package lotto.domain.model
 
 class WinningNumbers(val winningNumber: WinningNumber, val bonusNumber: BonusNumber)
