+++ src/main/kotlin/lotto/domain/service/WinningNumbersFactory.kt
@@ -1,4 +1,8 @@
-package lotto.domain
+package lotto.domain.service
+
+import lotto.domain.model.BonusNumber
+import lotto.domain.model.WinningNumber
+import lotto.domain.model.WinningNumbers
 
 class WinningNumbersFactory {
     fun create(winningNumbers: List<Int>, bonusNumber: Int): WinningNumbers {
