+++ src/test/kotlin/lotto/InputHandlerTest.kt
@@ -1,4 +1,4 @@
-import lotto.InputHandler
+import lotto.view.InputHandler
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.DisplayName
