+++ src/test/kotlin/lotto/domain/model/MoneyTest.kt
@@ -1,7 +1,7 @@
-package lotto.domain
+package lotto.domain.model
 
-import lotto.Messages.INPUT_NOT_DIVIDABLE
-import lotto.Messages.NUMBER_NOT_POSITIVE
+import lotto.config.Messages.INPUT_NOT_DIVIDABLE
+import lotto.config.Messages.NUMBER_NOT_POSITIVE
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.DisplayName
