+++ src/test/kotlin/lotto/domain/service/StatisticsTest.kt
@@ -1,5 +1,7 @@
-package lotto.domain
+package lotto.domain.service
 
+import lotto.domain.model.Money
+import lotto.domain.model.Rank
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
